src/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
python-firebase==1.2.1
requests==2.29.0
urllib3==1.26.15
